cleanup/package reorganisation
