+++ hazelcast-jet-core/src/main/java/com/hazelcast/jet/impl/metrics/CompressingProbeRenderer.java
@@ -14,18 +14,16 @@
  * limitations under the License.
  */
 
-package com.hazelcast.jet.impl.util;
+package com.hazelcast.jet.impl.metrics;
 
 import com.hazelcast.internal.metrics.renderers.ProbeRenderer;
 import com.hazelcast.nio.Bits;
 
-import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
-import java.io.DataInputStream;
 import java.io.DataOutputStream;
 import java.io.IOException;
 
-public class CompressingProbeRenderer implements ProbeRenderer {
+class CompressingProbeRenderer implements ProbeRenderer {
 
     public static final int TYPE_LONG = 0;
     public static final int TYPE_DOUBLE = 1;
@@ -103,28 +101,6 @@ public class CompressingProbeRenderer implements ProbeRenderer {
         } finally {
             dos = null;
             baos = null;
-        }
-    }
-
-    public static void decompress(byte[] data, ProbeRenderer renderer) {
-        DataInputStream dis = new DataInputStream(new ByteArrayInputStream(data));
-        String lastName = "";
-        try {
-            while (dis.available() > 0) {
-                int equalPrefixLen = dis.readUnsignedShort();
-                lastName = lastName.substring(0, equalPrefixLen)
-                        + dis.readUTF();
-                int type = dis.readByte();
-                if (type == TYPE_LONG) {
-                    renderer.renderLong(lastName, dis.readLong());
-                } else if (type == TYPE_DOUBLE) {
-                    renderer.renderDouble(lastName, dis.readDouble());
-                } else {
-                    throw new RuntimeException("Unexpected type");
-                }
-            }
-        } catch (IOException e) {
-            throw new RuntimeException(e); // can be thrown with unexpected EOFException
         }
     }
 }
